Linux環境に対応

+++ part3/ch28/28-4-4-has-not-equals/src/Vector.h
@@ -29,7 +29,7 @@ public:
 
 template<typename T>
 class Vector {
-	int sz;
+	size_t sz;
 	T* elem;
 public:
 	using value_type = T;
